SDDSIP-1101: Fix 500 error when going back after designated site removed (#494)
https://eaflood.atlassian.net/browse/SDDSIP-1101

As we go through the "Give details of conservation issues" flow, we enter details of a designated site. At the end, we get a "Check your answers" page with a "Remove site" option (`/designated-site-check-answers`). Removing the site and then using the Back link (or browser back button) gives us an error when we hit a page such as `/advice-from-natural-england`. Instead we simply want to be redirected back to `/designated-site-check-answers`.

We note that `/designated-site-remove` already includes a `checkDesignatedSite` function as part of its `checkData` which does exactly what we want: if checks to see if we are currently working on a designated site and if not, it redirects to `/designated-site-check-answers`. We therefore extract this out into `conservation-considerations/common.js` (and its tests into the matching `common.spec.js`) and add it to all relevant pages:

- `/designated-site-permission`
- `/details-of-permission`
- `/advice-from-natural-england`
- `/ne-activity-advice`
- `/designated-site-proximity`
