+++ packages/web-service/src/pages/conservation-considerations/designated-site-proximity.js
@@ -2,7 +2,7 @@ import Joi from 'joi'
 import pageRoute from '../../routes/page-route.js'
 import { conservationConsiderationURIs } from '../../uris.js'
 import { checkApplication } from '../common/check-application.js'
-import { allCompletion, getCurrentSite } from './common.js'
+import { allCompletion, getCurrentSite, checkDesignatedSite } from './common.js'
 import { APIRequests } from '../../services/api-requests.js'
 import { PowerPlatformKeys } from '@defra/wls-powerapps-keys'
 
@@ -23,7 +23,7 @@ export const setData = async request => {
 export default pageRoute({
   page: DESIGNATED_SITE_PROXIMITY.page,
   uri: DESIGNATED_SITE_PROXIMITY.uri,
-  checkData: checkApplication,
+  checkData: [checkApplication, checkDesignatedSite],
   validator: Joi.object({
     proximity: Joi.string().required().valid(...Object.values(PowerPlatformKeys.ON_SITE_OR_CLOSE_TO_SITE).map(v => v.toString()))
   }).options({ abortEarly: false, allowUnknown: true }),
